Replaced the old compare_bias.py with the new one.
